add check for url exists

+++ mgmicq/tasks/tasks.py
@@ -1,7 +1,7 @@
 from celery.task import task
 from time import sleep
 import paramiko as pk
-import os, json
+import os, json,httplib
 import getpass
 from dockertask import docker_task
 from subprocess import call,STDOUT
@@ -24,20 +24,18 @@ def mgmic_qc_workflow(forward_read_url, reverse_read_url, basedir="/data/static/
     task_id = str(mgmic_qc_workflow.request.id)
     resultDir = os.path.join(basedir, 'mgmic_tasks/', task_id)
     os.makedirs(resultDir)
+    #Check if Urls exist
+    if not check_url_exist(forward_read_url):
+        raise Exception("Please Check URL %s" % forward_read_url)
+    if not check_url_exist(reverse_read_url):
+        raise Exception("Please Check URL %s" % reverse_read_url)
     foward_read = os.path.join(resultDir,forward_read_url.split('/')[-1])
     reverse_read = os.path.join(resultDir,reverse_read_url.split('/')[-1])
     logfile= open(resultDir + "/logfile.txt","w")
     #get the forward and reverse read files
     #print 'wget','-O',foward_read,forward_read_url
-    try:
-        call(['wget','-O',foward_read,forward_read_url],stdout=logfile)
-    except:
-        raise Exception("Please Check URL %s" % forward_read_url)
-    #print 'wget','-O',reverse_read,reverse_read_url
-    try:
-        call(['wget','-O',reverse_read,reverse_read_url],stdout=logfile)
-    except:
-        raise Exception("Please Check URL %s" % reverse_read_url)
+    call(['wget','-O',foward_read,forward_read_url],stdout=logfile)
+    call(['wget','-O',reverse_read,reverse_read_url],stdout=logfile)
     logfile.close()
     docker_opts = "-v /opt/local/scripts/:/scripts -v /data/static:/data/static"
     docker_cmd = "/scripts/bin/Illumina_MySeq_Trim %s %s %s" % (foward_read,reverse_read,resultDir)
@@ -46,6 +44,14 @@ def mgmic_qc_workflow(forward_read_url, reverse_read_url, basedir="/data/static/
         return "http://%s/mgmic_tasks/%s" % (result['host'],result['task_id'])
     except:
         raise
+
+def check_url_exist(url):
+    c = httplib.HTTPConnection(url)
+    c.request("HEAD", '')
+    if c.getresponse().status == 200:
+        return True
+    else:
+        return False
 
 @task()
 def qc_docker_workflow(forward_read_filename, reverse_read_filename, basedir="/data/static/",docker_worker=os.environ['docker_worker']):
